pyproject.toml
[project]
name = "facepalm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "argon2-cffi>=25.1.0",
    "asgi-correlation-id>=4.3.4",
    "databases>=0.9.0",
    "fastapi>=0.121.1",
    "httpx>=0.28.1",
    "logtail-python>=0.3.4",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.11.0",
    "pytest>=9.0.0",
    "pytest-anyio>=0.0.0",
    "pytest-asyncio>=1.3.0",
    "python-jose>=3.5.0",
    "python-json-logger>=4.0.0",
    "python-multipart>=0.0.20",
    "rich>=14.2.0",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
[tool.pytest.ini_options]
pythonpath = ["facepalm"]
testpaths = ["src/tests"]
dependencies = [
    "databases[aiosqlite,asyncpg]>=0.9.0",
    "dotenv>=0.9.9",
    "fastapi>=0.119.1",
    "httpx>=0.28.1",
    "pydantic-settings>=2.11.0",
    "pytest>=8.4.2",
    "ruff>=0.14.1",
    "sqlalchemy>=2.0.44",
    "uvicorn[standard]>=0.38.0",
]
